Moved MapSelection class back to where it was.
