+++ bot/map_selection.py
